storage/requirements.txt
swagger_ui_bundle==1.1.0
connexion[flask,uvicorn]==3.1.0
Flask==2.3.3
SQLAlchemy==2.0.35
mysql-connector-python==9.0.0
PyMySQL==1.1.1
pykafka==2.8.0